Correct arguments for Rule class
Rules require three arguments:

* a left-hand side specifier (Predicate) that specifies a condition
that must be true for the rule to be a candidate to fire
* a right-hand-side key (String) that matches a key to be inferred
* a value (Expression) that specifies the value for the right-hand-side
key when a rule fires

In DirectToWeb’s Rule system, the right-hand-side is described in
Assignment objects, which contain both the key and the value. In this
system, the two parts of the right-hand-side are separated.